requirements.txt
# Core ML Libraries
torch>=2.1,<3.0
torchvision>=0.16,<0.20
mosaicml>=0.17.0

# Data and ML Utilities
datasets>=2.14.0
transformers>=4.30.0
numpy>=1.24.0
Pillow>=9.0.0

# Training and Monitoring
wandb>=0.15.0
tqdm>=4.64.0
matplotlib>=3.6.0
seaborn>=0.12.0

# HuggingFace Integration
huggingface_hub>=0.16.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0

# AWS (for deployment)
boto3>=1.28.0
awscli>=1.29.0
